backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
numpy==1.24.3
psycopg2-binary==2.9.6
deepface==0.0.79
easyocr==1.6.2
requests==2.28.2
opencv-python-headless==4.5.4.60
tkinterdnd2==0.3.0